Created tester application instead of running tests through Unity3d because of license activation issue

+++ Resources/TextureVoxelSurfaceShader.shader
@@ -40,16 +40,16 @@
 			float tileSizeUV = (float)_TileSize / (float)_TextureSize;
 			
 			// Convert from size/256
-			uint faceWidth = round(IN.uv2_WidthHeightIndex.x * 256.0);
-			uint faceHeight = round(IN.uv2_WidthHeightIndex.y * 256.0);
+			int faceWidth = round(IN.uv2_WidthHeightIndex.x * 256.0);
+			int faceHeight = round(IN.uv2_WidthHeightIndex.y * 256.0);
 			
 			// Covert from index/256
-			uint textureIndex = round(IN.uv3_TextureDetailIndex.x * 256.0);
-			uint detailIndex = round(IN.uv3_TextureDetailIndex.y * 256.0);
+			int textureIndex = round(IN.uv3_TextureDetailIndex.x * 256.0);
+			int detailIndex = round(IN.uv3_TextureDetailIndex.y * 256.0);
 			
-			uint tileNum = _TextureSize / _TileSize;
-			float2 textureIndexUV = float2(textureIndex % tileNum, (tileNum - 1) - textureIndex / tileNum) * tileSizeUV;
-			float2 detailIndexUV = float2(detailIndex % tileNum, (tileNum - 1) - detailIndex / tileNum) * tileSizeUV;
+			int tileNum = _TextureSize / _TileSize;
+			float2 textureIndexUV = float2((uint)textureIndex % (uint)tileNum, (tileNum - 1) - (uint)textureIndex / (uint)tileNum) * tileSizeUV;
+			float2 detailIndexUV = float2((uint)detailIndex % (uint)tileNum, (tileNum - 1) - (uint)detailIndex / (uint)tileNum) * tileSizeUV;
 			
 			float2 scaledUV = float2(IN.uv_TextureMap. x* faceWidth, IN.uv_TextureMap.y * faceHeight);
 			float2 currentUVBound = float2((uint)scaledUV.x % faceWidth, (uint)scaledUV.y % faceHeight);
